% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_brickset.R
\name{brickset_api}
\alias{brickset_api}
\title{Get API options}
\usage{
brickset_api(where, auth_args = list(key = brickset_key(), username =
  brickset_username(), password = brickset_password(), userHash =
  brickset_hash(), authtime = brickset_authtime()),
  default_args = list(), ...)
}
\arguments{
\item{where}{API endpoint (e.g. getTheme, getSets)}

\item{auth_args}{list containing key, username, password,
userHash (if known), and authtime (if known). If list is full of
NULL values (default), they will be substituted from the environment.
Environment variables should be named .brickset_key,
.brickset_username, .brickset_password, .brickset_authtime, and
.brickset_hash, though the last two should generally not be
user-specified.
Auth variables are only necessary if variables are not cached in the
global environment; if the global variables exist, this argument can
be ignored entirely.}

\item{default_args}{default arguments for the generic api. Should be supplied by calling function}

\item{...}{other API arguments}
}
\description{
This is a general helper for handling the brickset API.
}
