% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_brickset.R
\name{brickset_get_instructions}
\alias{brickset_get_instructions}
\title{Get link to instructions for a single lego set from brickset.com}
\usage{
brickset_get_instructions(setID, auth_args = list(key = brickset_key(),
  username = brickset_username(), password = brickset_password(), userHash
  = brickset_hash(), authtime = brickset_authtime()))
}
\arguments{
\item{setID}{Set ID, obtained from sets retrieved by brickset_get_sets().
This is NOT the set number, it's the internal unique database ID
that remains constant, whereas set numbers may not.}

\item{auth_args}{list containing key, username, password,
userHash (if known), and authtime (if known). If list is full of
NULL values (default), they will be substituted from the environment.
Environment variables should be named .brickset_key,
.brickset_username, .brickset_password, .brickset_authtime, and
.brickset_hash, though the last two should generally not be
user-specified.
Auth variables are only necessary if variables are not cached in the
global environment; if the global variables exist, this argument can
be ignored entirely.}
}
\description{
Get link to instructions for a single lego set from brickset.com
}
\examples{
\dontrun{
brickset_get_instructions(setID = 22941)
}
}
