% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_brickset.R
\name{brickset_get_themes}
\alias{brickset_get_themes}
\title{Get all lego themes from brickset.com}
\usage{
brickset_get_themes(auth_args = list(key = brickset_key(), username =
  brickset_username(), password = brickset_password(), userHash =
  brickset_hash(), authtime = brickset_authtime()))
}
\arguments{
\item{auth_args}{list containing key, username, password,
userHash (if known), and authtime (if known). If list is full of
NULL values (default), they will be substituted from the environment.
Environment variables should be named .brickset_key,
.brickset_username, .brickset_password, .brickset_authtime, and
.brickset_hash, though the last two should generally not be
user-specified.
Auth variables are only necessary if variables are not cached in the
global environment; if the global variables exist, this argument can
be ignored entirely.}
}
\description{
Get all lego themes from brickset.com
}
\examples{
\dontrun{
brickset_auth()
themes <- brickset_get_themes()
}
}
