% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_brickset.R
\name{brickset_setup}
\alias{brickset_setup}
\title{Set up R to work with brickset.com}
\usage{
brickset_setup()
}
\description{
Set up R to work with brickset.com
}
\examples{
\dontrun{
brickset_setup()
}
}
