% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pieces.R
\name{fit_lego}
\alias{fit_lego}
\title{Fit a linear model to predict the number of pieces based on year}
\usage{
fit_lego(lego_data, ...)
}
\arguments{
\item{lego_data}{A data frame or tibble containing LEGO set information, including "year" and "pieces" columns.}

\item{...}{Catches unused arguments.}
}
\value{
An object of class \code{"lego_fit"} which includes the model details
as well as the data set used for fitting.
}
\description{
Fits a linear regression model to predict the number of pieces in LEGO sets based on the year.
}
\note{
This function is specifically designed for LEGO set data containing "year" and "pieces" columns.
}
\examples{
mod <- fit_lego(lego_data)
print(mod)
}
