% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_desc.R
\docType{data}
\name{lego_data}
\alias{lego_data}
\title{LEGO Dataset}
\format{
A data frame with 2,669 observations on 14 variables.
\describe{
\item{name}{Name of the Lego set.}
\item{year}{Release year of the lego set.}
\item{theme}{Theme or brand of the Lego set.}
\item{themegroup}{The generic group of themes a Lego set belongs to.}
\item{pieces}{Number of pieces in a Lego set.}
\item{minifigs}{Number of miniature figurines included in the Lego set.}
\item{agerange_min}{The recommended minimum age for the Lego set.}
\item{us_retailprice}{The US retail price in US dollars at the time of the datasets creation.}
}
}
\usage{
lego_data
}
\description{
The data was taken from a dataset comprised of all lego sets from Kaggle.
}
\keyword{datasets}
