% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lego_store.R
\name{lego_get_set_data}
\alias{lego_get_set_data}
\title{Get data from a lego set page}
\usage{
lego_get_set_data(url)
}
\arguments{
\item{url}{page url}
}
\description{
Get data from a lego set page
}
\examples{
library(LegoR)
sets <- lego_get_sets("https://shop.lego.com/en-US/category/creator-3-in-1")
lego_get_set_data(sets$set_link[1]) # get one set
sets <- sets \%>\%
  dplyr::mutate(set_info = purrr::map(set_link, lego_get_set_data)) \%>\%
  tidyr::unnest(set_info)
}
