% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lego_store.R
\name{lego_get_sets}
\alias{lego_get_sets}
\title{Get all lego sets in a category}
\usage{
lego_get_sets(url)
}
\arguments{
\item{url}{url of the base page (additional pages will be retrieved automatically)}
}
\description{
Get all lego sets in a category
}
\examples{
library(LegoR)
lego_get_sets("https://shop.lego.com/en-US/category/creator-3-in-1")
}
