% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lego_store.R
\name{lego_get_sets_on_page}
\alias{lego_get_sets_on_page}
\title{Get all lego sets on the page}
\usage{
lego_get_sets_on_page(url)
}
\arguments{
\item{url}{url of page with lego sets}
}
\description{
Get all lego sets on the page
}
