% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lego_store.R
\name{lego_get_themes}
\alias{lego_get_themes}
\title{Get all lego themes from shop.lego.com}
\usage{
lego_get_themes(url = "https://shop.lego.com/en-US/category/themes")
}
\arguments{
\item{url}{URL of shop theme website (should also work for interests)}
}
\description{
Get all lego themes from shop.lego.com
}
\examples{
library(LegoR)
themes <- lego_get_themes()
interests <- lego_get_themes("https://shop.lego.com/en-US/category/interests")
}
