% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegoR-package.R
\docType{package}
\name{LegoR-package}
\alias{LegoR}
\alias{LegoR-package}
\title{LegoR: Package to scrape data from shop.lego.com and brickset}
\description{
Collects data from shop.lego.com and brickset for integration. The inst folder includes a script to automatically update downloaded data.
}
\author{
\strong{Maintainer}: Susan Vanderplas \email{srvanderplas@gmail.com} (0000-0002-3803-0972)

}
\keyword{internal}
