% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_api}
\alias{rebrickable_api}
\title{Generic API call}
\usage{
rebrickable_api(where, api_key = rebrickable_key(),
  default_args = list(), ..., follow_next = TRUE)
}
\arguments{
\item{where}{API endpoint}

\item{api_key}{API key (pulled from .rebrickable_key if it exists)}

\item{default_args}{default API args}

\item{...}{Additional arguments to the API}

\item{follow_next}{If response contains a 'next' argument, should it be followed?}
}
\description{
General helper for handling the brickset API.
}
