% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_api_all}
\alias{rebrickable_api_all}
\title{Get all pages from the rebrickable api}
\usage{
rebrickable_api_all(api_res)
}
\arguments{
\item{api_res}{rebrickable_api return object}
}
\value{
tibble of rebrickable_api return objects
}
\description{
Follows 'next' links returned by API calls, appending the results into
a tibble so that each tibble row corresponds to a single call.
}
