% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_parse_api_res}
\alias{rebrickable_parse_api_res}
\title{Function to format API response}
\usage{
rebrickable_parse_api_res(res)
}
\arguments{
\item{res}{response from API}
}
\value{
tibble with response status, content (list col), headers (list col), url, date, times (list col), and request (list col)
}
\description{
Function to format API response
}
