% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_part_category_info}
\alias{rebrickable_part_category_info}
\title{Get a data frame of information about a specific part category ID}
\usage{
rebrickable_part_category_info(cat_id = 1, key = rebrickable_key(),
  ..., parse = T)
}
\arguments{
\item{cat_id}{part category id}

\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
Get a data frame of information about a specific part category ID
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_part_category_info(cat_id = 1)
}
}
