% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_part_color_sets}
\alias{rebrickable_part_color_sets}
\title{Get a list of all sets a part/color combination has appeared in}
\usage{
rebrickable_part_color_sets(part_id = 1, color_id = 1,
  key = rebrickable_key(), ..., parse = T)
}
\arguments{
\item{part_id}{part ID}

\item{color_id}{color ID}

\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
Get a list of all sets a part/color combination has appeared in
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_part_color_sets(part_id = 31110, color_id = 1)
}
}
