% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_parts}
\alias{rebrickable_parts}
\title{Get a data frame of information about parts}
\usage{
rebrickable_parts(key = rebrickable_key(), ..., parse = T)
}
\arguments{
\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
The parts API endpoint takes parameters page, page_size, part_num, part_nums,
part_cat_id, color_id, bricklink_id, brickowl_id, legoid, ldraw_id, ordering, 
and search (a search term).
}
\details{
Note that currently external IDs are not returned.
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_parts(page = 1, follow_next = F)
}
}
