% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_set_parts}
\alias{rebrickable_set_parts}
\title{Get a data frame of information about pieces in a specific set}
\usage{
rebrickable_set_parts(set_id = 1, key = rebrickable_key(), ...,
  parse = T)
}
\arguments{
\item{set_id}{set ID}

\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
Note that external IDs are not preserved in parsed results.
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_set_parts(set_id = "40349-1")
}
}
