% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_sets}
\alias{rebrickable_sets}
\title{Get a data frame of information about sets}
\usage{
rebrickable_sets(key = rebrickable_key(), ..., parse = T)
}
\arguments{
\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
The parts API endpoint takes parameters page, page_size, theme_id, min_year,
max_year, min_parts, max_parts, ordering, and search.
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_sets(min_year = 2018, max_year = 2019, min_parts = 20, page_size = 1000, follow_next = T)
}
}
