% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_setup}
\alias{rebrickable_setup}
\title{Set up R to work with rebrickable.com}
\usage{
rebrickable_setup()
}
\description{
Set up R to work with rebrickable.com
}
\examples{
\dontrun{
rebrickable_setup()
}
}
