% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_themes}
\alias{rebrickable_themes}
\title{Get a data frame of all themes}
\usage{
rebrickable_themes(key = rebrickable_key(), ..., parse = T)
}
\arguments{
\item{key}{API key (pulled from environment if saved via rebrickable_save_credentials)}

\item{...}{other arguments (page, page_size, ordering)}

\item{parse}{Return results as a formatted tbl without the response information?}
}
\description{
Get a data frame of all themes
}
\examples{
if (exists(".rebrickable_key")) {
  rebrickable_themes()
}
}
