% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_brickset.R
\name{brickset_save_credentials}
\alias{brickset_save_credentials}
\title{Save brickset.com credentials to Rprofile file}
\usage{
brickset_save_credentials(username, password, api_key,
  profile_save = "user")
}
\arguments{
\item{username}{username.}

\item{password}{password.}

\item{api_key}{API key.}

\item{profile_save}{Save information to R profile? One of "user", "project", or NULL (to not save information)}
}
\description{
Opens the user's Rprofile file and pastes the values to the clipboard, then
sources the file.
}
\examples{
\dontrun{
# brickset_save_credentials("your_user", "your_password", "your_api_key")
}
}
