% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rebrickable.R
\name{rebrickable_save_credentials}
\alias{rebrickable_save_credentials}
\title{Save rebrickable credentials to Rprofile file}
\usage{
rebrickable_save_credentials(api_key, profile_save = "user",
  sys_env_var = "rebrickable_key", global_var = ".rebrickable_key")
}
\arguments{
\item{api_key}{API key.}

\item{profile_save}{Save information to R profile? One of "user", "project", or NULL (to not save information)}

\item{sys_env_var}{Variable name to use if saving system environment variable. Do not change - parameter for testing purposes only.}

\item{global_var}{Variable name to use if saving global variable. Do not change - parameter for testing purposes only.}
}
\description{
Opens the user's Rprofile file and pastes the values to the clipboard, then
sources the file.
}
\examples{
\dontrun{
# rebrickable_save_credentials("your_api_key")
}
}
